Add avatar synchronization and improve handling
Introduced server-client synchronization for avatar data using networking. Enhanced avatar fetching, storage, and cleanup logic, ensuring consistency across server and clients. Adjusted third-party SFS library initialization to prevent overwriting existing definitions.

+++ gamemode/core/libs/sh_avatars.lua
@@ -6,10 +6,49 @@ ow.avatars.stored = ow.avatars.stored or {}
 
 file.CreateDir("overwatch/avatars")
 
+if ( SERVER ) then
+    util.AddNetworkString("ow.avatars.fetch")
+else
+    net.Receive("ow.avatars.fetch", function()
+        local data = net.ReadString()
+        if ( !data ) then return end
+
+        local decoded, err = sfs.decode(data)
+        print(decoded, type(decoded))
+        if ( err ) then
+            ow.util:PrintError("ow.avatars.fetch", " Failed to decode avatar data: " .. err)
+            return
+        end
+
+        ow.avatars.stored = decoded
+
+        for sID, avatarData in pairs(ow.avatars.stored) do
+            if ( !istable(avatarData) or !isstring(avatarData.url) ) then continue end
+
+            file.Write("overwatch/avatars/" .. sID .. ".png", avatarData.data)
+        end
+    end)
+end
+
+if ( SERVER ) then
+    function ow.avatars:Sync()
+        local encoded, err = sfs.encode(self.stored)
+        if ( err ) then
+            ow.util:PrintError("ow.avatars:Sync", "Failed to encode avatar data: " .. err)
+            return
+        end
+
+        net.Start("ow.avatars.fetch")
+            net.WriteString(encoded)
+        net.Broadcast()
+    end
+end
+
 function ow.avatars:GetAvatar(ply)
     local steamID = ply:SteamID64()
-    if ( ow.avatars.stored[steamID] ) then
-        return ow.avatars.stored[steamID]
+    local stored = self.stored[steamID]
+    if ( stored ) then
+        return stored
     end
 
     http.Fetch(format:format("ISteamUser", "GetPlayerSummaries", STEAM_API_KEY, steamID), function(body, len, headers, code)
@@ -27,13 +66,17 @@ function ow.avatars:GetAvatar(ply)
         if ( !istable(playerData) or !isstring(playerData.avatarfull) ) then return end
 
         local avatarURL = playerData.avatarfull
-        ow.avatars.stored[steamID] = avatarURL
+        stored = avatarURL
         http.Fetch(avatarURL, function(imageData)
             file.Write("overwatch/avatars/" .. steamID .. ".png", imageData)
 
-            ow.avatars.stored[steamID] = {url = avatarURL, data = imageData}
+            stored = {url = avatarURL, data = imageData}
 
-            return ow.avatars.stored[steamID]
+            if ( SERVER ) then
+                self:Sync()
+            end
+
+            return stored
         end)
     end)
 end
@@ -44,6 +87,8 @@ function ow.avatars:Clear()
     for k, v in ipairs(file.Find("overwatch/avatars/*.png", "DATA")) do
         file.Delete("overwatch/avatars/" .. v)
     end
+
+    self:Sync()
 end
 
 function ow.avatars:Delete(ply)
@@ -51,34 +96,42 @@ function ow.avatars:Delete(ply)
 
     ow.avatars.stored[steamID] = nil
     file.Delete("overwatch/avatars/" .. steamID)
+
+    if ( SERVER ) then
+        self:Sync()
+    end
 end
 
-hook.Add("PlayerConnect", "ow.avatars.PlayerConnect", function(ply, ip)
-    ow.avatars:GetAvatar(ply)
-end)
+if ( SERVER ) then
+    local avatarReloaded = false
+    hook.Add("OnReloaded", "ow.avatars.OnReloaded", function()
+        if ( avatarReloaded ) then return end
+        avatarReloaded = true
 
-local avatarReloaded = false
-hook.Add("OnReloaded", "ow.avatars.OnReloaded", function()
-    if ( avatarReloaded ) then return end
-    avatarReloaded = true
-    ow.avatars:Clear()
+        for _, ply in player.Iterator() do
+            ow.avatars:GetAvatar(ply)
+        end
+    end)
 
-    for _, ply in player.Iterator() do
+    hook.Add("PlayerInitialSpawn", "ow.avatars.PlayerConnect", function(ply, ip)
         ow.avatars:GetAvatar(ply)
-    end
-end)
+    end)
 
-hook.Add("OnShutdown", "ow.avatars.OnShutdown", function()
-    ow.avatars.stored = {}
+    hook.Add("OnShutdown", "ow.avatars.OnShutdown", function()
+        ow.avatars.stored = {}
 
-    for _, sID64 in ipairs(file.Find("overwatch/avatars/*.png", "DATA")) do
-        file.Delete("overwatch/avatars/" .. sID64)
-    end
-end)
+        for _, sID64 in ipairs(file.Find("overwatch/avatars/*.png", "DATA")) do
+            file.Delete("overwatch/avatars/" .. sID64)
+        end
 
-hook.Add("PlayerDisconnected", "ow.avatars.PlayerDisconnected", function(ply)
-    local steamID = ply:SteamID64()
-    ow.avatars.stored[steamID] = nil
+        ow.avatars:Sync()
+    end)
 
-    file.Delete("overwatch/avatars/" .. steamID)
-end)
+    hook.Add("PlayerDisconnected", "ow.avatars.PlayerDisconnected", function(ply)
+        local steamID = ply:SteamID64()
+        ow.avatars.stored[steamID] = nil
+
+        file.Delete("overwatch/avatars/" .. steamID)
+        ow.avatars:Sync()
+    end)
+end
